Two corrections based on 1.5.0 code
1) `nft_get_info` does not require a `wallet_id`
2) `nft_mint_nft` uses `edition_total` _not_ `edition_count`

Of note: `the example response for `nft_get_nfts` uses `edition_count` and `edition_number`. The type declaration for `NFTInfo` in nft_info.py has `series_total` & `series_number`. So not sure if those are being remapped somewhere during serialization, is an oversight or a mistake in the example. It would make sense for it to be made consistent with the edition nomenclature.