Improve formatting of section 3.5: signage_points and infusion_points
For such a consensus critical section, formatting and style consistency of section 3.5 was very mixed.

This change improves formatting so that constants and mathematical expressions are consistently formatted as _italic_. Remark (quote) section formatting is fixed to include lists belonging to that remark.

One part (introduction of signage_point_iterations) was moved to a more logical place. Previously, it was in the midst of a paragraph with unrelated information.

Iterations usage has been unified to always use full numbers, and not full numbers in one place and numbers with a M suffix in another.

Typographic `’` have been replaced with plain `'`. (To ease editing and improve consistency.)