Migration of knowledgebase from zendesk
This pr includes a number of updates required to deprecate the knowledgebase on zendesk to be replaced with pages on the docs site.

FAQ - multiple updates to general Chia FAQ, farming, plotting, etc. Migrates all remaining knowledgebase articles to the necessary locations.

cat2-intro: updates to 1 FAQ to better align wording with the knowledgebase article and to make the question easier to search

sidebar: updates to ensure the CAT2 upgrade information is properly presented in the guide